+++ Diploma/Models/Operation.cs
@@ -9,18 +9,18 @@ namespace Diploma.Models
 {
     internal class Operation
     {
-        private readonly Int32 _id;
-        private readonly Int32 _idPosition;
+        private readonly Int64 _id;
+        private readonly Int64 _idPosition;
         private readonly String _name;
         private readonly String _description;
 
-        public Int32 Id => _id;
-        public Int32 IdPosition => _idPosition;
+        public Int64 Id => _id;
+        public Int64 IdPosition => _idPosition;
         public String Name => _name;
         public String Description => _description;
 
         // Конструктор со всеми полями
-        public Operation(Int32 id, Int32 idPosition, String name, String description)
+        public Operation(Int64 id, Int64 idPosition, String name, String description)
         {
             _id = id;
             _idPosition = idPosition;
@@ -30,16 +30,16 @@ namespace Diploma.Models
 
         public Operation()
         {
-            _id = (Int32)0;
-            _idPosition = (Int32)1003;
+            _id = (Int64)0;
+            _idPosition = (Int64)1003;
             _name = "Бездельничать";
             _description = "Просто ничего не делать";
         }
         // Создает объект Operation из SqlDataReader
         public static Operation FromDataReader(SqlDataReader reader)
         {
-            Int32 id = reader.GetInt32(reader.GetOrdinal("id"));
-            Int32 idPosition = reader.GetInt32(reader.GetOrdinal("Id_position"));
+            Int64 id = reader.GetInt64(reader.GetOrdinal("id"));
+            Int64 idPosition = reader.GetInt64(reader.GetOrdinal("Id_position"));
             String name = reader.GetString(reader.GetOrdinal("Name"));
             String description = reader.GetString(reader.GetOrdinal("Description"));
             return new Operation(id, idPosition, name, description);
